Add journey and aircraft type to output data
Aircraft value was just right there so that was easy to find.

The journey value consists of something like "YVR-LHR", and when the flight gets a proper flight number the flight number is put in callsign and the journey value is filled in as journey instead (this was sometimes previously undefined when it was in `planes[11]`). Not sure why it's output like that, but this always produces sensible results for me at least.